requirements.txt
# Web framework and server
fastapi==0.95.1
uvicorn==0.22.0
jinja2==3.1.2
python-multipart==0.0.6

# Data validation and models
pydantic==1.10.7

# Data processing
numpy==1.24.3
requests==2.31.0

# Database (if needed)
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Environment variables
python-dotenv==1.0.0

# Console formatting (for qa_system.py)
rich>=13.7.1

# Basic NLP (minimal requirements)
scikit-learn==1.3.2 